HBX-1992: Reorganize the reverse engineering strategy hierarchy - Rename 'org.hibernate.tool.internal.reveng.OverrideRepository' to 'org.hibernate.tool.internal.reveng.strategy.OverrideRepository'
