+++ orm/src/main/java/org/hibernate/tool/internal/reveng/strategy/OverrideRepository.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.reveng;
+package org.hibernate.tool.internal.reveng.strategy;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -29,8 +29,11 @@ import org.hibernate.tool.api.reveng.AssociationInfo;
 import org.hibernate.tool.api.reveng.RevengStrategy;
 import org.hibernate.tool.api.reveng.SchemaSelection;
 import org.hibernate.tool.api.reveng.TableIdentifier;
+import org.hibernate.tool.internal.reveng.MetaAttributeHelper;
+import org.hibernate.tool.internal.reveng.OverrideBinder;
+import org.hibernate.tool.internal.reveng.SQLTypeMapping;
+import org.hibernate.tool.internal.reveng.TableFilter;
 import org.hibernate.tool.internal.reveng.MetaAttributeHelper.SimpleMetaAttribute;
-import org.hibernate.tool.internal.reveng.strategy.DelegatingStrategy;
 import org.hibernate.tool.internal.util.JdbcToHibernateTypeHelper;
 import org.hibernate.tool.internal.util.TableNameQualifier;
 import org.jboss.logging.Logger;
